moved latticepoint and latticediameter from Analysis.jl to Lattice.jl since they provide generic functionality for any lattice.
